refactor -- moved files to more logical locations and added comments to code to make it easier to understand calculations using SQL
